history hookup

+++ tests/repo_test.rb
@@ -2,20 +2,16 @@
 
 # 
 
-require_relative "repository"
+require "repository.rb"
 require "minitest/autorun"
 
 #=============================================================================
 class TestRepository < Minitest::Test 
 
   def setup
-    
+    # placeholder in case we want to setup test env    
   end
-  # check to see if framework is working
-  def test_simple
-    assert_equal(4, 2+2)
-    assert_equal(6, 2*3) 
-  end
+  
 
   #----------------------------------------------------------------------
   def test_create
@@ -26,14 +22,14 @@ class TestRepository < Minitest::Test
     Dir.chdir('.test')
     Repository.create()
     assert(File.exist?('.repository'))
-    assert(File.exist?('.repository/.files'))
     assert(File.exist?('.repository/.stage'))
-    FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
   def test_status
+    # tests blank status
     if File.exist?('.test')
       FileUtils.rm_rf('.test')
     end
@@ -41,8 +37,23 @@ class TestRepository < Minitest::Test
     Dir.chdir('.test')
     Repository.create()
     Repository.status()
-    FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
+  end
+
+  #----------------------------------------------------------------------
+  def test_status2
+    # adds a file and checks if in status
+    if File.exist?('.test')
+      FileUtils.rm_rf('.test')
+    end
+    Dir.mkdir('.test')
+    Dir.chdir('.test')
+    Repository.create()
+    File.open('file1.txt', 'w') { |f| f.write("test text") }
+    Repository.status()
+    Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
@@ -57,8 +68,8 @@ class TestRepository < Minitest::Test
     a = ['.test_file1']
     Repository.add(a)      
     Repository.status()
-    #FileUtils.rm_rf('.test')
     Dir.chdir('..')
+    FileUtils.rm_rf('.test')
   end
 
   #----------------------------------------------------------------------
